+++ lua/settings/neovide.lua
@@ -1,10 +1,6 @@
 vim.g.neovide_remember_window_size = true
 vim.g.neovide_cursor_antialiasing = true
 vim.g.neovide_cursor_vfx_mode = 'wireframe'
-vim.g.neovide_transparency = 1
-vim.g.neovide_window_floating_opacity = 0
-vim.g.neovide_floating_blur = 0
-vim.g.neovide_window_floating_blur = 0
 
 vim.g.neovide_font_size_windows = 14
 vim.g.neovide_font_size_linux = 14