+++ lua/plugins/config/dressing_config.lua
@@ -1,0 +1,3 @@
+return function(use)
+	use({ 'stevearc/dressing.nvim', event = 'VimEnter' })
+end