+++ lua/plugins/config/bufdelete_config.lua
@@ -1,0 +1,25 @@
+return function(use)
+	use({
+		'famiu/bufdelete.nvim',
+		event = 'VimEnter',
+		config = function()
+			require('which-key').register({
+				['<leader>b'] = {
+					name = '+buffer',
+					D = {
+						function()
+							require('bufdelete').bufdelete(0, true)
+						end,
+						'Delete Current Buffer (Forced)',
+					},
+					d = {
+						function()
+							require('bufdelete').bufwipeout(0)
+						end,
+						'Delete Current Buffer',
+					},
+				},
+			})
+		end,
+	})
+end