+++ lua/plugins/config/lsp/init.lua
@@ -19,6 +19,14 @@ return function(use)
 				local hl = 'DiagnosticSign' .. type
 				vim.fn.sign_define(hl, { text = icon, texthl = hl, numhl = '' })
 			end
+
+			vim.lsp.handlers['textDocument/hover'] = vim.lsp.with(vim.lsp.handlers.hover, {
+				border = 'single',
+			})
+
+			vim.lsp.handlers['textDocument/signatureHelp'] = vim.lsp.with(vim.lsp.handlers.signature_help, {
+				border = 'single',
+			})
 		end,
 		config = function()
 			local prequire = require('personal.utils.prequire')