+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 
 ### Conf
+- **golang:** enabled dap config on golang files
 - **neovide:** neovide config blur is removed since they fixed the blur problem on telescope
 - **nvim_tree:** filtered node_modules and target for ts and rust respectively
 - **prepare.sh:** added prepare.sh