+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Feat
 - **comment.nvim:** removed kommentary in favor of comment.nvim
+- **dressing.nvim:** added dressing.nvim plugin
 - **null-ls:** added integration to curl and proselint
 - **telescope:** added custom picker to cwd to given folder