+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - **comment.nvim:** removed kommentary in favor of comment.nvim
 - **coq_nvim:** added coq_nvim
 - **dressing.nvim:** added dressing.nvim plugin
+- **lsp-display:** lsp display now shows borders
 - **null-ls:** added integration to curl and proselint
 - **telescope:** added custom picker to cwd to given folder