+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+
+<a name="v2.1.0"></a>
+## [v2.1.0] - 2021-12-16
 ### Chore
 - format
 
@@ -239,7 +242,8 @@
 - **indentation:** moved indentation to it's own folder
 
 
-[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v2.0.3...HEAD
+[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v2.1.0...HEAD
+[v2.1.0]: https://github.com/tigorlazuardi/nvim/compare/v2.0.3...v2.1.0
 [v2.0.3]: https://github.com/tigorlazuardi/nvim/compare/v1.3.1...v2.0.3
 [v1.3.1]: https://github.com/tigorlazuardi/nvim/compare/v1.3.0...v1.3.1
 [v1.3.0]: https://github.com/tigorlazuardi/nvim/compare/v1.2.0...v1.3.0